requirements2.txt
fastapi
uvicorn[standard]
sqlalchemy
google-genai
python-multipart
aiofiles
python-dotenv